requirements.txt
Flask==2.2.2
python-dotenv
Flask-PyMongo==2.3.0
pymongo==4.3.3
Werkzeug~=2.2.2
gunicorn